Fix dx find executions in public projects.
Previously, when running dx find jobs, we would query the server for all subjobs in a tree using their root job IDs, and without mentioning a project. This works fine, except in public projects (that the user has no permissions to otherwise), where the server disregards such projects and returns empty results for our query.

With the project mentioned by ID, the server searches through it, but this means we have to group the root IDs and query them for each project separately.