PTFM-11590 Make all dxfs operations wait until first init.
All nontrivial filesystem methods get a self.wait_until_initialized()
call at the beginning. This waits until the fs has been initialized for
the first time. To prevent deadlock, all filesystem methods that are
called in init are factored out into two methods foo and _foo, the
latter of which is for internal use and does not call
wait_until_initialized.