PTFM-11590 Another attempt to fix DXFS tests.
Run dx-mount without --foreground. I'm presuming and hoping that it
exits when and only when init() has finished running (and therefore the
filesystem is safe to use). When run with --foreground we would open a
subprocess and continue with the rest of the tests immediately. This
apparently allowed tests to begin before dx-mount had even attached to
the mountpoint.