Don't print ID of app(let) if we're going to --run it anyway.
If the user is going to call --run they're more likely interested in the
job ID than the executable ID. This change suppresses the executable ID
in that case, but in general, if there is a Makefile, "dx build ...
--run" will still spew out the output of the make process, making it
nontrivial to extract the job ID. So this isn't a general solution to
make the job ID easily available to the user, but it will at least fix
some simple tests.

(dx build --remote solves a similar problem by running the make step
separately from the upload step, so that the latter can be easily parsed
for the executable ID.)

This bug was exposed by changes made in 3f29d10. Previously it was
masked by the fact that we would print out the executable ID, then call
os.execv-- which transfers control without flushing buffers, so the
executable ID would never be seen by the user. The subprocess methods
do, however, flush the buffers and made the executable ID visible.