PTFM-12403 Tab completion fixes
- Update argcomplete dependency to pull in a copy that manages spaces after individual completions
- Remove space escaping logic that is no longer necessary (since argcomplete correctly escapes completions)
- Clean up a bit of dead code

This change also highlights the fact that the upload command in dx sh lacks local path completions and option completions. Code in DXCLICompleter (used by dx sh) is redundant with the completion hook logic in completers attached to the argparse parser, and with the completion driver logic in argcomplete. Work on eliminating this redundancy is blocked on https://github.com/kislyuk/argcomplete/issues/83 for now.