Hardcode python2.7 in the packaging script
This is necessary when packaging the SDK for CentOS 5 and 6 (in which
the default python is 2.6).

Also parameterize the target interpreter #! line in the packaging
script.

This reverts a portion of 5630a72 that incorporates #! line rewriting
into the python build target. It's generally inappropriate to do that
rewrite except when packaging for a specific target in a controlled
build environment, since we don't have enough information about which
interpreter installation is being used for the build, or which one
will be used in the target system, or the relationship between the
two.