PTFM-13049 Inject SDK path into sys.path in Setuptools entry points
Force our Python path to appear before everything else in
sys.path. This is necessary when an out-of-date installation of
Setuptools appears in sys.path, providing a version of
pkg_resources.py that is unable to read our Python package
requirements metadata, and crashes with a
pkg_resources.DistributionNotFound error.