PTFM-13157 First file download request blocks all subsequent requests
Summary:
If running on a worker, wait for the first file download chunk to come
back before issuing any more requests. This ensures that the backend can serve
cached data for the subsequent requests, instead of having all of the first
DXFILE_HTTP_THREADS requests simultaneously hit a cold cache.

Test Plan: verified that cache misses are reduced with the new client

Reviewers: dkillebrew, klai

Reviewed By: dkillebrew, klai

Differential Revision: https://phabricator.dnanexus.com/D108