dxpy.DXHTTPRequest: Escape from retry loop if timeout is exceeded
In addition to passing `timeout` through as a socket timeout, also use
it as a soft bound on time spent in the retry loop.

This is intended for use in synchronous code where getting stuck in an
indefinitely long wait due to a 503 mode on the server is the wrong
thing to do.

Reviewed in https://github.com/dnanexus/dx-toolkit/pull/84.