Unset PYTHONPATH when installing Python dependencies in virtualenv
This is necessary to be able to update pip and setuptools properly
inside the virtualenv, where we collect the dependencies. When
upgrading packages, pip will refuse to uninstall the old package in
the virtualenv if the package is also available from a path in
PYTHONPATH that is outside the virtualenv.