run("dx cd foo") in tests didn't actually have the desired effect
run would invoke dx within bash, which meant that the config would get
saved to the process config dir associated with the bash process. This
directory would be unavailable to other children of the original Python
process, so it would fall back to the user-global conf dir, which is
prone to interference by other processes.