PTFM-14387 Move job log message truncation
The truncation previously resided in the dx-log-stream wrapper, which
ingests stdout and stderr, but data sent directly to the log socket
bypasses it. The wrapper invokes the logger function, which is
responsible for forwarding the message. The logger function is used in
Python-based apps. This change eliminates confusing JSON parsing
errors when Python-based apps send log lines that are too long.