Rename app wizard template vars to avoid bogus compilation errors
This lengthens the variable names but has the advantage that the Python
template files now *superficially* appear to be (closer to) well-formed
Python. Longer term the right thing to do is probably to rename the
template files as *.py.template or something so tools aren't even
tempted to try and byte-compile them.