PTFM-15313 improving dxpy socket timeouts
Summary:
Modifying timeout handling in the core of the HTTP request call. Previously
we had a loophole, where 503 mode on the server would have a limited retry
mechanism. Another issue, is that most routes require high timeouts, while
the file download/upload require low socket timeouts.

The way this is handled, is to have a default high timeout of 10 minutes, and
the HTTP calls for file upload/download use a  60 second timeout.

Test Plan: existing tests

Reviewers: psung

Reviewed By: psung

Differential Revision: https://dnanexus.phacility.com/D370