Added script to slice zipped VCF
Adding a script to slice a block gzipped VCF file given the VCF file and
its tabix index.  Both the tabix index and the VCF file may be given as:
- a file name
- a URL
- a DXFile ID (note that project-<>:file-<> doesn't seem to work)

Also needed is a list of intervals which can be either a file, as above
or a comma separated list of intervals in the form chrom:start-stop.
These intervals are CLOSED, 1-based intervals.

The script will write the file to a provided filename (mandatory!), but
it can append to an existing file, and it also has an option to prevent
writing the last "empty" bgzip block signifying end of file.
Additionally, printing of the header is optional as well.

The script will decompress only the part of the VCF file needed to
determine variant inclusion in a file, so is particularly well-suited to
downloading entire chromosomes at a time.  On the other hand, this
script will perform very poorly when downloading a large number of small
regions (<1,000nt) very close together (<16,000nt separation).

Note that this script differs from tabix in that it does NOT look for an
END= tag in the INFO field to determine overlapping.  Additionally, it
does not look at the size of the REF allele to determine if the position
should be retrieved.  A variant is retrieved IF AND ONLY IF the reported
POS is contained in a target region.