PTFM-16030 Support proxy S3 API endpoints in UA
UA: Check to see if the url we are using for file uploads contains the
hostname of our transpacific TCP tunnel.  If it does, it will resolve
the IP of this hostname, swap out this hostname for the normal AWS
hostname, and then add an entry to curl's resolve table for the AWS
hostname and the TCP tunnel's IP address.  In this way, when we check
the certificate (which will ultimately come from AWS) the hostname
will match.