PTFM-15388 fixing failing upload test
Summary:
A previous upload test with upload permissions had required
the billTo to be lowered to UPLOAD permissions which is no longer
available due to PTFM-15388. Instead a new scoped token is created
to only allow UPLOAD permissions for projects and that is used
instead of the default token for this test.

Test Plan:
correct existing test in test_dxclient
TestDXClientUploadDownload.test_dx_upload_with_upload_perm

Reviewers: pnguyen

Reviewed By: pnguyen

Differential Revision: https://dnanexus.phacility.com/D509