Fix flaky test: destroy project with prejudice in teardown
The project destroy was observed to fail (due to jobs not in a terminal
state) after a test, even though the test terminated the workflow it
ran. My guess is that the termination was not completely processed and
left something in a "terminating" state. So destroy the project with
{terminateJobs: true} to prevent this from tripping us up.

(cherry picked from commit 779e19db9070c5736ebe10ac92c6066d11b3021d)