Fix connection state issue introduced by 384e8ef
With 384e8ef, dxpy.DXHTTPRequest(..., **kwargs) now passes
unrecognized kwargs to urllib3, not to requests. This causes
unrecognized kwargs to propagate all the way down to httplib, exposing
a way for httplib to end up in an inconsistent state (if it crashes
due to the unrecognized kwarg or for other reasons) and fail to serve
a subsequent request that wants to reuse the same connection.

This change abandons all the cached connections to eliminate this
possbility. It may be too heavy-handed, but the overhead of opening a
new connection is relatively low.