Close connection to prevent reaching limit on open files
Summary:
The HttpClient (which is held by the DXHTTPRequest object) cannot
reclaim the file handle it uses until the connection is closed (or until
the entire HttpClient is garbage collected). This patch makes us close
the connection as soon as the response is read. Leaving it open may
cause us to run into the limit on the number of open files if the
application is busy enough that GC of the HttpClient objects does not
happen in a timely manner.

Patch submitted by Oleg Nev

Test Plan: added a regression test

Reviewers: eworley

Reviewed By: eworley

Differential Revision: https://dnanexus.phacility.com/D583