Fail faster when user supplies path without project context
Summary:
Report a resolution error if the user supplies a path that is not
project-qualified and there is no project context. Previously this was
propagated back to the caller, which would typically then crash while trying to
read or write the specified path.

This simplification allows us to guarantee that every output from resolve_path
falls in one of a small number of easily described cases. These are now
formally specified in the docstring of resolve_path, but loosely speaking they
correspond to files specified by their path (requires a project), files
specified by ID (need not have a project specified), and JBORs.

I added some TODOs for future refactoring that should be done, too.

Test Plan: existing/updated unit tests

Reviewers: pnguyen

Reviewed By: pnguyen

Differential Revision: https://dnanexus.phacility.com/D601