PTFM-16656 Use flock to make dx-jobutil-add-output concurrency safe
Summary:
Prevents a race condition where another process can modify the file between
when we read the existing output spec and when we write the file back
out. Another race condition was also possible where we could read some prefix
of the JSON that some other process was in the middle of writing (the data that
was read would fail to parse).

Test Plan: added a regression test

Reviewers: eworley

Reviewed By: eworley

Subscribers: eworley

Differential Revision: https://dnanexus.phacility.com/D642