PTFM-16643 "dx build --app" takes a --region arg
Summary:
When building an app using dx build --app, the script creates a
temporary project in which the previous applet will be stored. Originally,
this would have just taken the defaultRegion of the user, however it would
be much easier if the user could specify a region in "dx build --app".
Therefore a --region argument was added to dx build to specify the region
of the app.

Test Plan: test_dxclient.TestDXBuildApp.test_build_app_with_region

Reviewers: psung

Reviewed By: psung

Differential Revision: https://dnanexus.phacility.com/D650