Fix behavior of dx download on dxlink syntax after 5197bf60
We previously incorrectly identified the syntax

  '{"$dnanexus_link": "file-X"}'

as indicating an explicit project. Thus it would fail when supplied as
an argument to "dx download" (this usage is common in applications that
just do "dx download $myinputvar" without parsing the link first).

This change factors out the dict-parsing logic from resolve_path and
uses it in is_project_explicit as well.