Install dxpy via wheel to avoid using pkg_resources (fixup for 87d6040)
With 87d6040, we no longer distribute setuptools with dxpy. The
executable script entry points (console_scripts) were still using the
pkg_resources API provided by setuptools, exposing us to incompatible
old versions of setuptools or complete absence of that module on the
host. By first compiling a wheel and then installing from it, we
induce a new executable script entry point installation format, which
doesn't use the pkg_resources API at all.

A side effect of this is that it's no longer easy to identify such a
script and inject our "please source the environment file" help
message, so it's removed until we can figure out how to re-incorporate
it.