PTFM-17064 make 'dx get' work with open source apps
Summary:
It was desired that 'dx get' work with open source apps, since we were telling customers to
perform such atrocities as

  dx get $(dx describe app-foo --json | jq -r .applet)

This is now done directly on apps with 'openSource' field set to true, and we may all sleep
peacefully knowing that balance has been restored to the Force.

Test Plan:
Create an app with openSource == True and one with
openSource == False and run 'dx get' with both. Examine the
behavior; in the first case it should be as though we ran get
on an applet, in the second case it should fail.

Relevant test:

test_dxclient.TestDXBuildApp.test_get_app

Reviewers: psung, sking

Reviewed By: psung, sking

Differential Revision: https://dnanexus.phacility.com/D724