Followup to thread safety
Summary:
Removing unnecessary special handling of the first file chunk in the
download case . This is now not required, because we lock the cached
download url whenever it changes.

Note: My remaining concern is with the class methods in the DXFile
class; these have to do with starting up the thread pool. I think
there are some potential problems there.

Test Plan: dxpy integration test

Reviewers: psung

Reviewed By: psung

Differential Revision: https://dnanexus.phacility.com/D772