PTFM-15849 Allow disabling Retry Logic for all 5xx errors
Summary:
Currently, a 503 status code deploys a retry strategy for a number of
iterations. Here, an environment variable is created so that the user has the
option to disable this retry strategy and propagate the 503 code back to the
caller. The retry strategy may also be disabled for other 500 errors.

Test Plan:
DXHTTPRequestTest.java contains the tests in which an environment
(that disables the retry strategy) is fed into a DXHTTPRequest.

Reviewers: psung

Reviewed By: psung

Differential Revision: https://dnanexus.phacility.com/D750