PTFM-17201 Remove default download duration
Summary:
With PTFM-16904 we would like to throw an error whenever duration
is specified with a non-preauthenticated url since these download
urls no longer expire with a download token but rather with the
client's own token. However, dx-toolkit always specified a default
duration as well as to not preauthenticate urls. This change removes
that default duration for the url.

Test Plan: added regression test

Reviewers: brogoff, psung

Reviewed By: psung

Differential Revision: https://dnanexus.phacility.com/D777