SCI-424 "assetDepends" metadata to the dxapp.json
Summary:
An asset is a tarball that contains executables, libraries, static data or any other artifact that is required to run an app/applet.
The asset tarball is stored in the platform as a hidden file which in turn is referenced by a visible 'record' of type asset.
An app/applet will refer to an asset with runSpec.assetDepends metadata in either of the two ways as shown below:

```
"assetDepends": [
  {
    "name": "asset_bundle_record_name",
    "project": "asset_bundle_record_project_id",
    "folder": "folder_name",
    "version": "asset_bundle_record_version_number"
  },
  {
    "id": "asset_bundle_record_id"
  }
]
```

During app/applet build time the assetDepends i.e. the record object will be translated into a 'bundledDepends' object.

Each asset bundle record object contains following metadata:

```
--type AssetBundle
--property version=0.0.1
--details {"tarballId": "hidden_asset_bundle_tarball_file_id"}
```

Test Plan:
Added test cases to TestDXBuildApp suite to cover these functionalities.

./src/python/test/test_dxclient.py TestDXBuildApp

Reviewers: psung

Reviewed By: psung

Subscribers: jdale

Differential Revision: https://dnanexus.phacility.com/D781