dxfile.read: handling incomplete read error
Summary:
It is possible to get an incomplete read exception through the dxfile.read route.
This can happen if the data source is slow, and does not send a buffer full
of data within the timeout. This fix retries that case, with a smaller buffer.

Fixing lint error

Test Plan: The Scale test is a know way to reproduce this failure

Reviewers: psung

Reviewed By: psung

Subscribers: cjiang

Differential Revision: https://dnanexus.phacility.com/D877