Better recovery from partial reads
Summary:
An HTTP GET request can return partial data, if the socket
is closed before all the data arrives. This can happen due to a timeout,
and is a rare failure. Previously, we would discard the partial data, and
start over. This patch harvests the received data, and does a GET for the
rest of it.

Test Plan: Error injection

Reviewers: psung

Reviewed By: psung

Subscribers: cjiang

Differential Revision: https://dnanexus.phacility.com/D883