PTFM-17451 Use project context in get_download_url
Summary:
When get_download_url is called use the project id contained in the
DXFile object. If the api fails because the file in not in the project, then
retry the operation without a project hint.

Test Plan:
Added a small test in test_dxpy.py TestDXFile. Other tests in that
suite cover the behavior as well.

Reviewers: sking

Reviewed By: sking

Subscribers: psung

Differential Revision: https://dnanexus.phacility.com/D853