Added ability to dereference symlinks
Adding the ability to dereference symbolic links that are pointing to
remote files (i.e. not in the resource directory) of an app.  This patch
will ensure the following cases:

1. Links to local files and directories are kept
2. Links to remote files are dereferenced
3. Broken links cause an AppBuilderError
4. Links to remote directories cause an AppBuilderError
5. Links to local links (regardless of destination) are kept
6. Links to remote links (links to files) are dereferenced
7. Absolute links are treated as remote links (case 2,4)
8. Links that travel outside of the resource directory (i.e. ../resources/foo)
are treated as remote links (case 2,4).  This is pretty evil; please
don't be evil.

In any case, adding the new option "--force-symlinks" to the "dx build"
command will restore the previous behavior and keep symlinks that may or
may not resolve.  The use case for this could be perhaps a link in
resources that will point to a directory known to exist in a DNAnexus
instance (like /home/dnanexus for example).

Note that in the upload_resources_advanced test relies on the "old"
behavior of symlinks, so I also added the "--force-symlinks" tag to
those unit tests.