Recovery from partial reads by reducing buffer size
Summary:
There is an error case where we download a file, and get an IncompleteRead
error. This occurs when some of the data arrives, but not all of it.
It is not possible to harvest the partial data with urllib3, although,
it was possible with urllib2.

This change goes back to an older recovery strategy, that reduces the buffer size.
It does not try to piece together message fragments.

Test Plan: Existing tests

Reviewers: psung

Reviewed By: psung

Subscribers: cjiang

Differential Revision: https://dnanexus.phacility.com/D908