Updating download help message
To resume an interupted download, somewhat confusingly we use the -f or --overwrite flag.
I have updated the help message for this switch to specifically mention that it will resume
the file download if possible.  Long term we may want to either rename this option or separate
the two behaviors, but at least for now this adds some clarity.