Updated branch incorporating comments
Incorporating comments from @psung.  Specifically:
- Refactored the check for a local link into its own function.  This
  should correct the inappropriate error noted
- Added the test case for building with a relative path
- Replaced command-line tar invokation with the python tarfile library
- Cleaned up some comments and added inline documentation
- Adding tests for changing symlinks without --force-symlinks

Note that I did not make the mtime in the hash calculation the maximum
of the link mtime and the target's mtime.  The reason for this is if the
name of a link changes (but the target stays the same), the resource bundle
would still be rebuilt because the filename changes.  If the target changes,
the mtime would change, causing a resource bundle rebuild.  Additionally,
changing any link within the resource directory would cause the mtime of the
containing directory to change, also forcing a rebuild.