Fix problematic umask
When building a resource bundle, a umask that does not contain "o+r"
can be problematic when unpacking the bundle.  This fix changes the
permissions on the fly for any file or directory as it is being uploaded
as if the user were to "chmod -R a+rX".  Note that the checksum uses
the updated permissions in the calculation, so if permissions were
to change that would not affect the resultant tarball, the resource
bundle would not be built.