PTFM 17407: Add CLI for org-xxxx/findApps
Summary: dx toolkit interface to org-xxxx/findApps

Test Plan:
Create some apps billed to an org, and some that aren't. Check that the
ones returned from the CLI match the ones that are billed to the org. Much
of the testing was already done in the test for the API route already.

Reviewers: pnguyen

Reviewed By: pnguyen

Subscribers: tnguyen

Differential Revision: https://dnanexus.phacility.com/D867