PTFM-17766 Fix for NamedTemporaryFile on Windows
Summary:
The issue with NamedTemporaryFile on Windows is that the test was
opening multiple handles to the same file, which is not allowed in python.
To fix this I implemented a wrapper class around the NamedTemporaryFile,
which returns a file-like object in either closed or open state, so that it can
be written to subsequently.

Test Plan:
The test in test_dxpy.py TestDXFILE.test_download_project_selection.
Also the tests in test_dxclient.TestDXClientUploadDownload

Reviewers: psung, #systems

Reviewed By: psung, #systems

Subscribers: arina, psung, sking

Differential Revision: https://dnanexus.phacility.com/D940