Verify that 'less' works on Windows
Summary:
If dx runs on a windows machine that does not have gnu tools in the PATH,
then the pager fails and prints a message saying that less is not installed.
The fix is to test is 'less' is recognized on windows before invoking it.

Test Plan:
test_dxclient.TestDXClient.test_windows_pager Added this test
to remove the MinGW tools from the path before calling dx and asserting that the
error message is not printed on the output.

Reviewers: bagee

Reviewed By: bagee

Subscribers: bagee

Differential Revision: https://dnanexus.phacility.com/D1001