Handle ClosedPoolError during DXHTTPRequest
DXHTTPRequest uses a global urllib3.PoolManager to actually make requests. The
PoolManager manages urllib3.ConnectionPools, reusing them when appropriate. If
a request raises an exception, the PoolManager is "cleared", which closes
every ConnectionPool it manages. In rare cases, a ConnectionPool can be
assigned to a request but then closed before the request is made. This causes
a urllib3.exceptions.ClosedPoolError to be raised, which presently is not
handled and causes the DXHTTPRequest to fail.

The ideal solution would fix the concurrency issue, but it appears to be
within urllib3 itself. So, this change just catches ClosedPoolError and
retries.