Remove legacy dx-toolkit-{beta,unstable} packages
Summary:
These packages were once intended to be used to allow the use of
staging/master builds of dx-toolkit. Since then we decided that the
actual way to do this was to host pre-release versions of the packages
in separate APT repositories entirely. This change stops creating new
versions of the -beta and -unstable packages. In the unlikely event that
anyone is still using them in their app/let, it will continue to work
but they will henceforth just receive the version of dx-toolkit that
happens to be current at the time that this change is pushed.

Test Plan: none

Reviewers: pnguyen

Reviewed By: pnguyen

Differential Revision: https://dnanexus.phacility.com/D1009