Fix dx-download-all-inputs tests
Summary:
Previously when we had wanted to test "bash helper" scripts, we would
upload them as resources to use and then run test applets as theses uploaded
scripts would take precedence over the preinstalled scripts. However,
an upgrade on dx-download-all-inputs added a dependency on dxpy. Since master
and staging dxpy are out of sync, the stock staging dxpy does not contain the
necessary functions for the new dx-download-all-inputs.

To fix this, we load the master dx-toolkit into the worker rather than just
the few scripts and their dependencies. We then make the repo and source it
to make it the default dx-toolkit install. This will ensure that any test
that relies on changes to dxpy will obtain the correct version of dxpy in staging.

Test Plan: current tests

Reviewers: psung

Reviewed By: psung

Differential Revision: https://dnanexus.phacility.com/D1089