PTFM-18587 - Initial dx-docker script for dx-toolkit
Summary: The dx-docker script and its dependencies allows a user to run Docker containers within DNAnexus applications.  It is meant to be a seed commit that ensures that dependencies are properly installed and includes basic features of 'run' and 'pull' commands.  Additional features that we have developed and plan to develop will be iteratively incorporated into dx-tooklit as required.

Test Plan: In addition to manual testing, we have also designed a DNAnexus application that encodes a bunch of assertions of basic functionality. These tests will be continuously run after a seed commit of dx-docker is a part of dx-toolkit.

Reviewers: psung

Reviewed By: psung

Subscribers: knafissi, mlin

Differential Revision: https://dnanexus.phacility.com/D1137