Remove unreliable test
In the case where the tasks are supposedly kicked off
simultaneously (do_first_task_sequentially=False), the tasks can be
arbitrarily delayed due to system load. If that happens, it's impossible
for the test to distinguish between that and the case where the code
waits for the first task to finish. The test is useless if it can't
distinguish those scenarios, so I removed it entirely.