PTFM-18471 Use python's tarfile to extract applet's resources
Summary:
Calling tar as a subprocess on Windows may fail if the user does not
have the gnu tools installed. The fix is to use python's tarfile library to do
the extraction of resources when calling 'dx get'

Test Plan:
Added a test to test_dxclient.TestDXGetExecutables.test_get_applet_on_windows
that overrides the Windows 'PATH' to ignore gnu tools and make sure that the command 'dx get' succeeds.

Reviewers: gduggal

Reviewed By: gduggal

Subscribers: bagee

Differential Revision: https://dnanexus.phacility.com/D1133