PTFM-18776: Beta test dx-docker script: incremental improvements
Summary: This commit makes a few improvements to the dx-docker script so that it works overall more robustly in different situations.  It also prints metadata about the image to STDERR when executing `docker run`.  This metadata can be suppressed using the `-q` option to the subcommand.

Test Plan: This has been tested to work using the initial test suite made for dx-docker.

Reviewers: psung

Reviewed By: psung

Subscribers: mlin

Differential Revision: https://dnanexus.phacility.com/D1160