PTFM-18850 Remove "resources" field from app optional fields in "dx get".
Summary:
Currently, we cannot "dx get" an app and then "dx build --app" another app
(after making the necessary changes to actually build another app, like
changing the "version" field in the dxapp.json) because of the "resources"
field in the dxapp.json. After executing "dx get" on an app, the resulting
dxapp.json will contain a "resources" field set to the *resources container id*
of the app. When "dx build --app" is executed, the "resources" field specified
in the data of the /app/new request is set to the value of the corresponding
field in dxapp.json (a *container id*). Unfortunately, the "request" field in
the data of the /app/new request must be a string (*project id*) or an array of
strings (data object ids).

Because the "resources" field (the *container id*) is not needed to construct
the local app dir, we do not need to write it to the dxapp.json as part of the
"dx get" utility.

CHANGELOG: Fix bug to allow app developers to "dx get" an app and then "dx
build --app" a different version, for example, of the same app.

Test Plan:
Will add unit and integration test after verifying that this is the
right fix.

Reviewers: psung

Reviewed By: psung

Differential Revision: https://dnanexus.phacility.com/D1167