PTFM-18961 Remove installer dependency on pypi
Summary:
Changelog: This change allows the Windows CLI installer to run without
Internet access. Previously, access to pypi was required at install
time, so that pip could fetch the Python packages that dxpy depends on.
We now bundle a .whl file for each dxpy dependency with the installer.

Test Plan:
Built and ran the CLI installer on Win7, and verified that
authentication and uninstallation still behave normally.

Reviewers: nicolasbockg

Reviewed By: nicolasbockg

Differential Revision: https://dnanexus.phacility.com/D1203