PTFM-18962 dxpy not supplying user agent since 384e8ef
Summary: This diff fixes an issue where the user agent was not being passed by dxpy calls after a conversion to use urllib3.  This solution merges the headers populated in two different locations and uses them in requests.

Test Plan: A simple 'system/headers' API route was added to return headers of a request and the toolkit tests now contain a simple verification that the user agent exists and begins with dxpy.  All client tests were run as well as fast tests for the API server.

Reviewers: psung

Differential Revision: https://dnanexus.phacility.com/D1206