PTFM-18692: Do not merge headers. In some cases a bug is introduced.
Summary: In certain cases (e.g. file upload), the Authorization header meant for the apiserver is being sent to Amazon which results in an error (thanks to sking for noticing this). The solution is to not merge the headers and simply add to the '_headers' variable at the time the request is being sent.

Test Plan: Ran all client tests successfully. Running a branch test as a sanity check as well.

Reviewers: sking

Reviewed By: sking

Differential Revision: https://dnanexus.phacility.com/D1213