PTFM-18767 Limit number of threads in dx-download-all-inputs
Summary:
Using dx-download-all-inputs --parallel can cause instances with a small amount of
RAM to run out of memory. This commit limit the number of threads that download-all-inputs
can use so that there is at least 1200 MB of memory per thread. This is an empiric limit
based on the current minimum of 868 MB in mem1_hdd2_x8 instance types.

Test Plan:
Added a test TestDXBashHelpers.limit_threads that builds an applet that
uses dx-download-all-inputs on a mem1_hdd2_x8 instance and verifies that the number
of threads is indeed reduced from 8 to 5.

Reviewers: psung

Reviewed By: psung

Differential Revision: https://dnanexus.phacility.com/D1241