PTFM-18562 Run make in test runner on non-Mac OSes
Summary:
As of b9d8487, we stopped running 'make python src_libs' inside
run_python_integration_tests.py, since it was causing problems on OS X,
where dx-toolkit is now assumed to be running in a virtualenv.

As of this change we allow the make command to run again on non-Mac hosts.

Test Plan: Ran test runner script

Reviewers: psung

Reviewed By: psung

Subscribers: kjensen

Differential Revision: https://dnanexus.phacility.com/D1255